+++ index_investment/Portfolios/fornybarenergiindex.csv
