+++ index_investment/portfolio.py
@@ -1,8 +1,8 @@
 import os
 
-from Portfolio.Spiders.manage_threads import ManageThreads
-from Portfolio.config import ENCODING
-from Portfolio.config import PORTFOLIOS
+from index_investment.Spiders.manage_threads import ManageThreads
+from index_investment.config import ENCODING
+from index_investment.config import PORTFOLIOS
 
 
 def read_file(filename):
